split data.js to data/...ts files
