+++ components/charts/RadarChart.tsx
@@ -7,18 +7,23 @@ import {
    RadarChart,
    Tooltip
 } from 'recharts';
-import { skills } from '@/app/data';
+import { skills } from '@/data/skills';
 
-const sumRatingsByStackArea = skills.reduce((acc: Record<string, { sumOfRating: number; count: number }>, skill) => {
-   const { stackArea, rating } = skill;
-   if (acc[stackArea]) {
-      acc[stackArea].sumOfRating += rating;
-      acc[stackArea].count += 1;
-   } else {
-      acc[stackArea] = { sumOfRating: rating, count: 1 };
-   }
-   return acc;
-}, {});
+const sumRatingsByStackArea = skills.reduce(
+   (acc: Record<string, { sumOfRating: number; count: number }>, skill) => {
+      const { stackArea, rating } = skill;
+      if (stackArea === undefined || rating === undefined) return acc;
+
+      if (acc[stackArea]) {
+         acc[stackArea].sumOfRating += rating;
+         acc[stackArea].count += 1;
+      } else {
+         acc[stackArea] = { sumOfRating: rating, count: 1 };
+      }
+      return acc;
+   },
+   {}
+);
 
 const stackAreaRatingsArray = Object.keys(sumRatingsByStackArea).map(
    stackArea => ({
@@ -32,7 +37,11 @@ const stackAreaRatingsArray = Object.keys(sumRatingsByStackArea).map(
 );
 
 const MyRadarChart = () => (
-   <ResponsiveContainer width="100%" height="100%" minWidth={300} minHeight={300}>
+   <ResponsiveContainer
+      width="100%"
+      height="100%"
+      minWidth={300}
+      minHeight={300}>
       <RadarChart
          outerRadius={90}
          // width={730}
