+++ data/about.ts
@@ -1,0 +1,78 @@
+export const about = {
+   firstName: 'Bennett',
+   lastName: 'Brooks',
+   preferredName: 'Ben',
+   title: 'Software Engineer',
+   otherTitles: ['Web Developer', 'Full-Stack Developer', 'Frontend Developer'],
+   location: 'Portland, Oregon',
+   summary: 'Experienced software engineer',
+   overview:
+      'Experienced and versatile software engineer with a front-end leaning, full-stack background and a passion for continuous improvement. Known for clear communication, empathy, and a big-picture mindset, I thrive as a dedicated team player and contributor. With leadership experience and a love for mentoring, I bring a growth-oriented approach to any engineering role, always eager to learn, support others, and deliver impactful solutions.'
+};
+
+export const expertise = [
+   'Front-End + Full-Stack',
+   'Responsive Design',
+   'Accessibility / WCAG',
+   'Pixel-Perfect Implementation',
+   'Performance Optimization',
+   'Microservices',
+   'API Design & Development',
+   'A/B Testing & Experimentation',
+   'Monitoring & Observability',
+   'Process Improvement & Technical Leadership',
+   'Mentoring & Team Development',
+   'Cloud Infrastructure & CI/CD',
+   'Agile / Scrum'
+];
+
+export const education = [
+   {
+      degree: 'Bachelor of Science',
+      major: 'Mathematics',
+      institution: 'University of Redlands',
+      location: 'Redlands, CA',
+      startYear: 1996,
+      endYear: 2001,
+      awards: [
+         'Endowed Alumni Scholarship',
+         'Presidential Scholarship',
+         'Scholar-Athlete'
+      ]
+   },
+   {
+      degree: 'Master-Level Coursework',
+      major: 'Applied Mathematics',
+      institution: 'Claremont Graduate University',
+      location: 'Claremont, CA',
+      startYear: 2001,
+      endYear: 2001,
+      awards: [
+         'Fully Paid PhD Fellowship',
+         'Excelled, tutored peers, led study groups',
+         'Focused on fluid dynamics and differential equations',
+         'Participated in sponsored workshops to optimize semiconductor performance',
+         'Helped distinguished professors emeritus produce the second edition of their popular textbook on differential equations. Devised complex mathematical problems and authored solutions. Edited and proofread the text. Used LaTeX to typeset the book. Trained other students in LaTeX. Developed a love of tech. Also enhanced my life-long love of mentoring and teaching.',
+         "Wasn't planning on grad school but couldn't pass up the fully-paid PhD fellowship. Although I enjoyed the work, I realized that I didn't want to be a mathematician. I followed my new-found love of software to a career in tech."
+      ]
+   }
+];
+
+export const certifications = [
+   {
+      name: 'LaunchDarkly Gold Certification',
+      date: '2024'
+   },
+   {
+      name: 'Software Security Training',
+      date: '2021-2024'
+   },
+   {
+      name: 'Information Security Awareness (AgLearn)',
+      date: '2020'
+   },
+   {
+      name: 'Nielsen Norman Group User Experience training',
+      date: '2012'
+   }
+];
