+++ data/experience.ts
@@ -1,0 +1,470 @@
+export const experience = [
+   {
+      role: 'Contract Web Developer',
+      company: 'Self-Employed',
+      location: 'Portland, OR',
+      start: '2024-12',
+      end: '',
+      duties: [
+         {
+            name: '',
+            description:
+               'Building fast, beautiful websites and apps using JavaScript, React, Next.js, Vercel, MongoDB, Node, CSS Modules, and Tailwind CSS. Focused on clean architecture, maintainable code, performance, usability, and accessibility.',
+            skillNames: [
+               'JavaScript',
+               'TypeScript',
+               'Node.js',
+               'React.js',
+               'Next.js',
+               'Redux',
+               'Vercel',
+               'CSS Modules',
+               'CSS',
+               'Tailwind',
+               'RESTful APIs',
+               'Fastify',
+               'MongoDB',
+               'Zustand'
+            ]
+         }
+      ]
+   },
+   {
+      role: 'Staff Full-Stack Software Engineer',
+      company: 'ZenBusiness',
+      location: 'Austin, TX (Remote)',
+      start: '2021-02',
+      end: '2024-12',
+      duties: [
+         {
+            name: 'UI/UX Development & Integration',
+            description:
+               'Built business-critical apps, components, libraries, and design systems using JavaScript, React, Redux, Vue.js, Vuex, CSS Modules, and Styled Components.',
+            moreInfo: [
+               [
+                  'Built real-time, interactive applications with complex state management.',
+                  'Developed and maintained apps, components, libraries, and design systems, enhancing user experience (UX), developer experience (DX), and performance.',
+                  'Implemented responsive design and accessibility best practices.',
+                  'Contributed to the development of customer-facing applications, including onboarding flows, shopping/cards, dashboards, post-purchase/upsells and more.',
+                  'Contributed to the development of internal tools and applications.',
+                  'Improved and maintained the company’s marketing website using PHP, WordPress, JavaScript, CSS, Fastly, Cypress, and SEO, significantly improving core web vitals and coding standards.',
+                  'Led key improvements to performance, accessibility, and usability.',
+                  'Led efforts to improve performance and coding practices for multiple React-based repos.',
+                  'Developed and maintained a design system for consistent UI across products.',
+                  'Collaborated with design teams to implement engaging, pixel-perfect designs.',
+                  'Collaborated with product engineers, UI/UX designers and other cross-functional stakeholders.',
+                  'Developed atomic component libraries for consistent and efficient development.',
+                  'Integrated UI with APIs and data sources.',
+                  'Mentored junior and senior engineers and provided technical guidance.',
+                  'Defined technical requirements, collaborated with Product and Design teams, and led story mapping sessions.',
+                  'Improved coding and logging standards, ensuring clean, maintainable code across projects.',
+                  'Agile / Scrum methodologies, including story mapping, sprint planning, retrospectives, and daily stand-ups.'
+               ]
+            ],
+            skillNames: [
+               'JavaScript',
+               'TypeScript',
+               'Node.js',
+               'React.js',
+               'Vue.js',
+               'Redux',
+               'CSS',
+               'Sass/SCSS',
+               'CSS Modules',
+               'Styled Components',
+               'Jest',
+               'Cypress',
+               'RESTful APIs',
+               'A/B Testing',
+               'Multivariate Experimentation',
+               'LaunchDarkly',
+               'PostgreSQL',
+               'MongoDB',
+               'GraphQL',
+               'Micro-frontends'
+            ]
+         },
+         {
+            name: 'Web Development & Optimization',
+            description:
+               "Enhanced the company's public website using WordPress, PHP, SCSS/SASS, and JavaScript. Improved core web vitals and coding standards. Built a component library to increase efficiency and consistency.",
+            moreInfo: [
+               [
+                  'Developed, implemented and documented a component library for consistent UI across the site.',
+                  'Supported the shift to a page-speed-based SEO strategy, collaborating with the SEO Team to meet performance targets (e.g. TTFB, LCP, CLS).',
+                  'Vastly improved core web vitals and coding standards by making systematic optimization and recording performance metrics.',
+                  'Provided documentation, training and support related to maintaining performance and coding best practices.',
+                  'Planned, implemented, documented and supported an atomic component structure, facilitating quick and consistent development.',
+                  'Mentored junior and senior engineers and provided technical guidance.',
+                  'Built automated tests using Cypress.',
+                  'Agile / Scrum methodologies, including story mapping, sprint planning, retrospectives, and daily stand-ups.'
+               ]
+            ],
+            // moreInfoLink: {
+            //    href: '/experience/zb-seo',
+            //    text: 'Read more about optimizing the website...'
+            // },
+            skillNames: [
+               'WordPress',
+               'PHP',
+               'Sass/SCSS',
+               'CSS Modules',
+               'CSS',
+               'JavaScript',
+               'Responsive Design',
+               'SEO',
+               'Performance Optimization',
+               'A/B Testing',
+               'Multivariate Experimentation',
+               'LaunchDarkly'
+            ]
+         },
+         {
+            name: 'Microservices & APIs',
+            description:
+               'Built and maintained scalable microservices to enhance functionality and replace monolithic APIs. Used Node.js, Fastify, JavaScript, GCP, Kubernetes, and Docker to develop microservices and APIs.',
+            moreInfo: [
+               [
+                  'Built and maintained scalable microservices to improve developer experience.',
+                  'Leveraged technologies like Node.js, Fastify, JavaScript, GCP...',
+                  'Supported the decommissioning of a monolithic API.',
+                  'Contributed to the development of Node.js/Fastify microservices.',
+                  'Developed robust solutions for validating JSON and XML data, and automatically converting between data formats.',
+                  'Led the development of a critical microservice to automate the gathering, storage, preparation, validation and submission of data for a federal API, boosting efficiency and revenue.',
+                  'Defined requirements, collaborated with stakeholders, led story mapping and development. ',
+                  'Integrated logging and monitoring in GCP to ensure system reliability.',
+                  'Helped improve team coding and logging standards.',
+                  'Built unit tests using Jest.',
+                  'Provided technical support across codebases.',
+                  'Mentored senior and junior engineers.',
+                  'Implemented secure storage and processing of Personally Identifiable Information (PII) using a proxy to handle sensitive data.',
+                  'Provided technical guidance throughout the development lifecycle and represented the team in cross-functional meetings.',
+                  'Managed projects from technical scoping to deployment and user support, ensuring delivery of high-quality features aligned with business goals.',
+                  'Agile / Scrum methodologies, including story mapping, sprint planning, retrospectives, and daily stand-ups.'
+               ]
+            ],
+            skillNames: [
+               'JavaScript',
+               'Node.js',
+               'Fastify',
+               'Python',
+               'Jest',
+               'Microservices',
+               'RESTful APIs',
+               'GCP',
+               'GKE',
+               'AWS',
+               'Docker',
+               'Colima',
+               'Kubernetes',
+               'Terraform',
+               'Django',
+               'Micro-frontends'
+            ]
+         },
+         {
+            name: 'Experimentation',
+            description:
+               'Led a team dedicated to experimentation and optimization, focusing on A/B testing and multivariate experimentation. Helped build a culture of experimentation to speed iterative improvements and increase revenue.',
+            moreInfo: [
+               [
+                  'Led planning, building, tracking, and reporting on A/B and multivariate tests.',
+                  'Collaborated with Design and Product teams to develop hypotheses, define metrics, and design variants.',
+                  'Built variants.',
+                  'Worked with Data teams to analyze and present results.',
+                  'Partnered with stakeholders to confirm winners, implement solutions, and plan follow-up tests.',
+                  'Mentored senior engineers on testing and coding practices.',
+                  'Facilitated the transition to LaunchDarkly as our needs became too complex for Google Optimize.',
+                  'Trained other teams and presented to stakeholders regarding experimentation and LaunchDarkly.',
+                  'Helped build a culture of experimentation to speed iterative improvements and increase revenue.',
+                  'Acquired LaunchDarkly Gold Certification and provided ongoing support.',
+                  'Agile / Scrum methodologies, including story mapping, sprint planning, retrospectives, and daily stand-ups.'
+               ]
+            ],
+            skillNames: [
+               'A/B Testing',
+               'Multivariate Experimentation',
+               'LaunchDarkly',
+               'JavaScript',
+               'React.js',
+               'Vue.js',
+               'PHP',
+               'CSS',
+               'Sass/SCSS',
+               'Styled Components',
+               'Responsive Design'
+            ]
+         },
+         {
+            name: 'Developer Support & DevOps',
+            description:
+               "Focused on supporting developers and improving developer experience, gained basic DevOps knowledge and experience. While my passions lie closer to the frontend, it's extremely valuable to understand the backend and DevOps.",
+            moreInfo: [
+               [
+                  'Provided technical support and guidance to developers.',
+                  'Administered and documented developer tools.',
+                  'Contributed to API abstractions.',
+                  'Contributed to release workflows and CI/CD.',
+                  'Contributed to a centralized event distributor using Pub/Sub.',
+                  'Agile / Scrum methodologies, including story mapping, sprint planning, retrospectives, and daily stand-ups.'
+               ]
+            ],
+            skillNames: [
+               'ArgoCD',
+               'GCP',
+               'Kubernetes',
+               'Pub/Sub',
+               'Auth0',
+               'Twilio',
+               'Mailchimp',
+               'JavaScript',
+               'Node.js',
+               'RESTful APIs',
+               'Docker',
+               'Colima',
+               'Helm',
+               'Terraform'
+            ]
+         }
+      ],
+      skills: [
+         'JavaScript',
+         'TypeScript',
+         'React.js',
+         'Vue.js',
+         'Next.js',
+         'Redux',
+         'Thunk',
+         'CSS',
+         'SCSS',
+         'Sass',
+         'Styled Components',
+         'Responsive Design',
+         'A/B and multivariate testing',
+         'LaunchDarkly',
+         'Google Optimize',
+         'Webpack',
+         'Turbopack',
+         'Parcel',
+         'Jest',
+         'Cypress',
+         'Node.js',
+         'PHP',
+         'WordPress',
+         'Python',
+         'Express',
+         'Fastify',
+         'Django',
+         'MongoDB',
+         'PostgreSQL',
+         'SQL',
+         'RESTful APIs',
+         'GraphQL',
+         'Microservices',
+         'micro-frontends',
+         'GCP',
+         'GKE',
+         'AWS',
+         'Docker',
+         'Colima',
+         'Kubernetes',
+         'Helm',
+         'Terraform'
+      ]
+   },
+   {
+      role: 'Owner & Lead Engineer',
+      company: 'NuTandem',
+      location: 'Portland, OR',
+      start: '2013-05',
+      end: '2021-12',
+      duties: [
+         {
+            name: 'App Development',
+            description:
+               'Developed hundreds of custom apps and websites using JavaScript, React, Node, d3.js, CSS Modules, Styled Components, and Tailwind CSS',
+            moreInfo: [
+               [
+                  'Built a library of reusable, stylable, composable components to speed delivery and facililtate consistent usability, accessability and performance.',
+                  'Oversaw operations. Owned projects from requirements to delivery.'
+               ]
+            ],
+            skillNames: [
+               'JavaScript',
+               'TypeScript',
+               'Node.js',
+               'React.js',
+               'Redux',
+               'D3.js',
+               'Sass/SCSS',
+               'CSS Modules',
+               'Styled Components',
+               'Tailwind',
+               'CSS',
+               'Responsive Design',
+               'RESTful APIs',
+               'GraphQL',
+               'PostgreSQL',
+               'AWS'
+            ]
+         },
+         {
+            name: 'Web Development, SEO & Optimization',
+            description:
+               'Built our company website with WordPress, PHP, JavaScript, and CSS. Optimized SEO and performance.',
+            moreInfo: [
+               [
+                  'Designed and built our company website with WordPress, PHP, JavaScript, and CSS, including custom templates, menus, and widgets, to facilitate frequent updates.',
+                  'Led SEO, marketing and sales efforts.',
+                  'Managed the company’s online presence and digital marketing strategy.',
+                  'Implemented SEO best practices to improve search engine rankings and drive organic traffic.'
+               ]
+            ],
+            skillNames: [
+               'Next.js',
+               'WordPress',
+               'PHP',
+               'JavaScript',
+               'CSS Modules',
+               'CSS',
+               'SEO',
+               'Performance Optimization',
+               'Responsive Design'
+            ]
+         },
+         {
+            name: 'Everything Else',
+            description:
+               'Established and maintained relationships with over 50 organizations including state and federal agencies, restaurants like Sizzler and Outback, universities, and large corporations like Unilever.',
+            moreInfo: [
+               [
+                  'Managed all aspects of the business, including relationships, marketing, and customer support.',
+                  'Maintained solid relationships with large corporations (like Massage Envy, Sizzler Restaurants, Outback Steakhouse, and Unilever), government agencies (such as the USDA Forest Service and multiple state agencies), universities (including University of Montana, Central Oregon Community College, and Carroll University) and a variety of other customers.',
+                  'Provided training and support to clients.',
+                  'Owned projects from discovery to delivery.',
+                  'Conducted usability studies and worked with clients to understand their unique needs.'
+               ]
+            ]
+         }
+      ]
+   },
+   {
+      role: 'Contract Web Developer',
+      company: 'USDA Forest Service',
+      location: 'Albuquerque, NM (Remote)',
+      start: '2013-06',
+      end: '2020-12',
+      duties: [
+         {
+            name: 'Web Development & Performance Optimization',
+            description:
+               'Managed web projects and improved system performance using JavaScript, CSS, PHP and SQLite.',
+            moreInfo: [
+               [
+                  'Developed and maintained a custom web application for the USDA Forest Service.',
+                  'Juggled a steady stream of web projects for internal websites and the public website at www.fs.usda.gov',
+                  'Built web apps with JavaScript, CSS, PHP, and SQLite.',
+                  'Reduced reliance on jQuery by implementing pure JavaScript components.',
+                  'Built templates to enable self-serve website creation for non-technical users.',
+                  'Used Git and GitHub to share, review, and manage code. Participated in code reviews and incremental refactoring.',
+                  'Helped demonstrate app enhancements and performance improvements to leadership.',
+                  "Lead a project to build a Content Review Tool to support the agency's cumbersome efforts to review and improve hundreds of thousands of pages in their  public website.",
+                  'Agile / Scrum methodologies, including story mapping, sprint planning, retrospectives, and daily stand-ups.'
+               ]
+            ],
+            skillNames: [
+               'JavaScript',
+               'Node.js',
+               'PHP',
+               'SQLite',
+               'SQL',
+               'CSS',
+               'CSS3',
+               'Responsive Design',
+               'Performance Optimization',
+               'RESTful APIs'
+            ]
+         },
+         {
+            name: 'Web Accessibility & Training Expert',
+            description:
+               'Trained colleagues, conducted code reviews, and managed web projects. Built web apps using JavaScript, CSS, PHP and SQLite.',
+            moreInfo: [
+               [
+                  'Trained with a senior web accessbility expert (WCAG).',
+                  'Improved web accessibility for Forest Service websites and templates.',
+                  'Designed and delivered WCAG training.',
+                  'Provided training related to self-serve website creation.',
+                  'Provided ongoing technical support.'
+               ]
+            ],
+            skillNames: ['Accessibility (WCAG)']
+         }
+      ]
+   },
+   {
+      role: 'Systems Manager & Process Engineer',
+      company: 'GL Solutions',
+      location: 'Bend, OR',
+      start: '2006-07',
+      end: '2013-05',
+      duties: [
+         {
+            name: 'Systems Manager & Process Engineer',
+            description:
+               'Re-engineered processes and tools across the organization. Improved efficiency, accuracy, customer satisfaction and employee engagement, and usability and adoption for internal software tools. Defined SCAS, KPIs, and team metrics. Implemented a continuous improvement initiative, empowering employees to actively participate in evolving the organization. Provided custom software, documentation, training, and support.',
+            moreInfo: [
+               [
+                  'Re-engineered processes and tools across the organization.',
+                  'Improved efficiency, accuracy, customer satisfaction and employee engagement.',
+                  'Defined SCAS, KPIs, and team metrics.',
+                  'Implemented a continuous improvement initiative, empowering employees to actively participate in evolving the organization.',
+                  'Provided custom software, documentation, training, and support.',
+                  'Integrated customer service, sales, and marketing teams into a cohesive Business Solutions Department.',
+                  'Re-engineered processes and tools to better utilize talent and experience.',
+                  'Worked directly with customers to increase engagement.',
+                  'Facilitated idea sharing and built a customer community.',
+                  'Conducted stakeholder and SME interviews, documented requirements, and produced software specifications.',
+                  'Authored and maintained technical content, and improved documentation standards.',
+                  'Wrote detailed responses to government RFPs, RFQs and RFIs. Maintained an in-depth library of content for use in formal responses. Responded to questions and provided support to sales to close deals and transition new customers.',
+                  'Conducted stakeholder and SME interviews, documented requirements, and produced software specifications. Authored and maintained technical content, and improved documentation standards.',
+                  'Agile / Scrum methodologies, including story mapping, sprint planning, retrospectives, and daily stand-ups.'
+               ]
+            ],
+            skillNames: [
+               'SQL',
+               'CSS',
+               'HTML5',
+               'Responsive Design',
+               'Usability Testing',
+               'User Experience (UX)',
+               'User Interface (UI)',
+               'Agile Development',
+               'Scrum'
+            ]
+         }
+      ]
+   },
+   {
+      role: 'Technical Documentation Specialist',
+      company: 'I/O Software',
+      location: 'Riverside, CA',
+      start: '2002-03',
+      end: '2006-07',
+      duties: [
+         {
+            name: '',
+            description:
+               'Produced technical documentation for end users and internal users. Contributed to patent applications related to cryptographic security.',
+            moreInfo: [
+               [
+                  'Produced user guides, in-app help systems, and technical articles related to software integrations with biometric security devices (e.g. fingerprint, retina, iris, voice, face, and keystroke).',
+                  'Interviewed developers and other stakeholders.',
+                  'Contributed to cryptographic patent applications related to security algorithms and biometric data mappings, including research, mathematical consultation and diagrams.'
+               ]
+            ],
+            skillNames: ['CSS', 'HTML5']
+         }
+      ]
+   }
+];
