+++ data/projects.ts
@@ -1,1223 +1,6 @@
 import { getYearsSince } from '@/utils/date';
+import { DefinitionProject, ProjectImage } from '@/types/Project';
 
-export const about = {
-   firstName: 'Bennett',
-   lastName: 'Brooks',
-   preferredName: 'Ben',
-   title: 'Software Engineer',
-   otherTitles: ['Web Developer', 'Full-Stack Developer', 'Frontend Developer'],
-   location: 'Portland, Oregon',
-   summary: 'Experienced software engineer',
-   overview:
-      'Experienced and versatile software engineer with a front-end leaning, full-stack background and a passion for continuous improvement. Known for clear communication, empathy, and a big-picture mindset, I thrive as a dedicated team player and contributor. With leadership experience and a love for mentoring, I bring a growth-oriented approach to any engineering role, always eager to learn, support others, and deliver impactful solutions.'
-};
-export const expertise = [
-   'Front-End + Full-Stack',
-   'Responsive Design',
-   'Scalable & Accessible UI',
-   'Pixel-Perfect Implementation',
-   'SEO & Performance Optimization',
-   'Microservices & APIs',
-   'A/B Testing & Experimentation',
-   'Monitoring & Observability',
-   'Process Improvement & Technical Leadership',
-   'Mentoring & Team Development',
-   'Cloud Infrastructure & CI/CD'
-];
-export const skills = [
-   {
-      type: 'Language',
-      stackArea: 'General',
-      stackCategory: 'Language',
-      name: 'JavaScript',
-      icon: '/js.svg',
-      cloudinary: 'js_au9kmj',
-      cloudinaryDark: 'javascript-dark-icon_diwklq',
-      yearLearned: 2013,
-      years: 12,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Frontend Library',
-      stackArea: 'Frontend',
-      stackCategory: 'Library',
-      name: 'React.js',
-      icon: '/react.svg',
-      cloudinary: 'react_g4rwdx',
-      cloudinaryDark: 'react-dark-icon_dibccl',
-      yearLearned: 2015,
-      years: 9,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Frontend Framework',
-      stackArea: 'Frontend',
-      stackCategory: 'Framework',
-      name: 'Next.js',
-      icon: '/next.svg',
-      cloudinary: 'nextjs_jxdbqq',
-      cloudinaryDark: 'next-dark-icon_qfvtkk',
-      yearLearned: 2019,
-      years: 4,
-      rating: 9,
-      haveUsedProfessionally: false,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Frontend Framework',
-      stackArea: 'Frontend',
-      stackCategory: 'Framework',
-      name: 'Vue.js',
-      icon: '/vue.svg',
-      cloudinary: 'vue_thuhnk',
-      cloudinaryDark: 'vue-dark-icon_xghpvp',
-      yearLearned: 2021,
-      years: 3,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Backend Development',
-      stackArea: 'Backend',
-      stackCategory: 'Language',
-      name: 'Node.js',
-      icon: '/node.svg',
-      cloudinary: 'node_hl58om',
-      cloudinaryDark: 'node-dark-icon_kl7oy6',
-      yearLearned: 2015,
-      years: 9,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Language',
-      stackArea: 'General',
-      stackCategory: 'Language',
-      name: 'TypeScript',
-      icon: '/ts1.svg',
-      cloudinary: 'ts1_nbdvxm',
-      cloudinaryDark: 'typescript-dark-icon_mj51g8',
-      yearLearned: 2015,
-      years: 9,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'State Management',
-      stackArea: 'Frontend',
-      stackCategory: 'Library',
-      name: 'Redux',
-      icon: '/redux.svg',
-      cloudinary: 'redux_vd28jd',
-      yearLearned: 2019,
-      years: 8,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'State Management',
-      stackArea: 'Frontend',
-      stackCategory: 'Library',
-      name: 'Zustand',
-      icon: '/react.svg',
-      cloudinary: 'zustand_wrx1q3',
-      yearLearned: 2024,
-      years: 1,
-      rating: 8,
-      haveUsedProfessionally: false,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Visualization',
-      stackArea: 'Frontend',
-      stackCategory: 'Library',
-      name: 'D3.js',
-      cloudinary: 'd3js_eakt6f',
-      yearLearned: 2013,
-      years: 8,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Visualization',
-      stackArea: 'Frontend',
-      stackCategory: 'Library',
-      name: 'React Simple Maps',
-      cloudinary: 'react-simple-maps_xvvujp',
-      yearLearned: 2013,
-      years: 8,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Styling',
-      stackArea: 'Frontend',
-      stackCategory: 'Styling',
-      name: 'CSS',
-      icon: '/css3.svg',
-      cloudinary: 'css3_mvgbbf',
-      yearLearned: 2004,
-      years: 12,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Styling',
-      stackArea: 'Frontend',
-      stackCategory: 'Styling',
-      name: 'CSS Modules',
-      icon: '/sass.svg',
-      cloudinary: 'css3_mvgbbf',
-      cloudinaryDark: 'css3-dark-icon_xtdb3t',
-      yearLearned: 2015,
-      years: 8,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Styling',
-      stackArea: 'Frontend',
-      stackCategory: 'Styling',
-      name: 'Sass/SCSS',
-      icon: '/sass.svg',
-      cloudinary: 'sass_s3kz3x',
-      cloudinaryDark: 'sass-dark-icon_h3esvt',
-      yearLearned: 2015,
-      years: 8,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Styling',
-      stackArea: 'Frontend',
-      stackCategory: 'Styling',
-      name: 'Tailwind',
-      icon: '/tailwind.svg',
-      cloudinary: 'tailwind_k9r34s',
-      cloudinaryDark: 'tailwind-dark-icon_mbvvtr',
-      yearLearned: 2019,
-      years: 4,
-      rating: 9,
-      haveUsedProfessionally: false,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Styling',
-      stackArea: 'Frontend',
-      stackCategory: 'Styling',
-      name: 'JSS',
-      icon: '/jss.svg',
-      cloudinary: 'jss_zewece',
-      yearLearned: 2016,
-      years: 8,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: false
-   },
-   {
-      type: 'Styling',
-      stackArea: 'Frontend',
-      stackCategory: 'Styling',
-      name: 'Styled Components',
-      icon: '/styled-components.svg',
-      cloudinary: 'styled-components_fd8fvo',
-      cloudinaryDark: 'styled-components-dark-icon_fs0kmr',
-      yearLearned: 2016,
-      years: 8,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Styling',
-      stackArea: 'Frontend',
-      stackCategory: 'Styling',
-      name: 'Panda CSS',
-      icon: '/react.svg',
-      cloudinary: 'panda-css_av1hkt',
-      yearLearned: 2025,
-      years: 1,
-      rating: 9,
-      haveUsedProfessionally: false,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: false
-   },
-   {
-      type: 'UI/UX',
-      stackArea: 'Frontend',
-      stackCategory: 'UI/UX',
-      name: 'Responsive Design',
-      icon: '/rwd.svg',
-      cloudinary: 'rwd_o9z8s0',
-      yearLearned: 2007,
-      years: 11,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'UI/UX',
-      stackArea: 'Experimentation & Optiminzation',
-      stackCategory: '',
-      name: 'A/B Testing',
-      icon: '/beaker.svg',
-      cloudinary: 'beaker_vqshpg',
-      yearLearned: 2010,
-      years: 3,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'UI/UX',
-      stackArea: 'Experimentation & Optiminzation',
-      stackCategory: '',
-      name: 'Multivariate Experimentation',
-      icon: '/beaker1.svg',
-      cloudinary: 'beakers_yellow_eehkpx',
-      yearLearned: 2022,
-      years: 3,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'UI/UX',
-      stackArea: 'Experimentation & Optiminzation',
-      stackCategory: '',
-      name: 'LaunchDarkly',
-      icon: '/launchdarkly.svg',
-      cloudinary: 'launchdarkly_kdznir',
-      yearLearned: 2022,
-      years: 3,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Build Tools',
-      stackArea: 'Build & Testing Tools',
-      stackCategory: 'Build Tools',
-      name: 'Webpack',
-      icon: '/webpack.svg',
-      cloudinary: 'webpack_ycdnkv',
-      yearLearned: 2015,
-      years: 7,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Build Tools',
-      stackArea: 'Build & Testing Tools',
-      stackCategory: 'Build Tools',
-      name: 'Turbopack',
-      icon: '/turbopack.svg',
-      cloudinary: 'turbopack_gvkvbu',
-      yearLearned: 2024,
-      years: 1,
-      rating: 7,
-      haveUsedProfessionally: false,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Build Tools',
-      stackArea: 'Build & Testing Tools',
-      stackCategory: 'Build Tools',
-      name: 'Parcel',
-      icon: '/parcel.svg',
-      cloudinary: 'parcel_ik2cum',
-      yearLearned: 2024,
-      years: 1,
-      rating: 7,
-      haveUsedProfessionally: false,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: false
-   },
-   {
-      type: 'Testing',
-      stackArea: 'Build & Testing Tools',
-      stackCategory: 'Testing',
-      name: 'Jest',
-      icon: '/jest.svg',
-      cloudinary: 'jest_djvbcr',
-      yearLearned: 2022,
-      years: 3,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Testing',
-      stackArea: 'Build & Testing Tools',
-      stackCategory: 'Testing',
-      name: 'Cypress',
-      icon: '/cypress.svg',
-      cloudinary: 'cypress_ihahr3',
-      yearLearned: 2021,
-      years: 1,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Backend Development',
-      stackArea: 'Backend',
-      stackCategory: 'Language',
-      name: 'PHP',
-      icon: '/php.svg',
-      cloudinary: 'php-icon_apfbnf',
-      cloudinaryDark: 'php01-dark-icon_fzefry',
-      yearLearned: 2013,
-      years: 9,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Backend Development',
-      stackArea: 'General',
-      stackCategory: 'Content Management',
-      name: 'WordPress',
-      icon: '/wordpress.svg',
-      cloudinary: 'wordpress_xkgfcd',
-      cloudinaryDark: 'wordpress-dark-icon_volzbh',
-      yearLearned: 2013,
-      years: 9,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Backend Development',
-      stackArea: 'Backend',
-      stackCategory: 'Language',
-      name: 'Python',
-      icon: '/python.svg',
-      cloudinary: 'python_g5utim',
-      cloudinaryDark: 'python-dark-icon_ogismc',
-      yearLearned: 2021,
-      years: 3,
-      rating: 7,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: true,
-      showInSkillsTable: true,
-      featureInCTAs: true
-   },
-   {
-      type: 'Backend Framework',
-      stackArea: 'Backend',
-      stackCategory: 'Framework',
-      name: 'Express',
-      icon: '/express.svg',
-      cloudinary: 'express_bqzgin',
-      yearLearned: 2015,
-      years: 9,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Backend Framework',
-      stackArea: 'Backend',
-      stackCategory: 'Framework',
-      name: 'Fastify',
-      icon: '/fastify.svg',
-      cloudinary: 'fastify_zy8jul',
-      yearLearned: 2023,
-      years: 2,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Backend Framework',
-      stackArea: 'Backend',
-      stackCategory: 'Framework',
-      name: 'Django',
-      icon: '/django.svg',
-      cloudinary: 'django_yilgyq',
-      yearLearned: 2020,
-      years: 2,
-      rating: 4,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Database',
-      stackArea: 'Data Storage & Databases',
-      stackCategory: 'NoSQL',
-      name: 'MongoDB',
-      icon: '/mongodb.svg',
-      cloudinary: 'mongodb_kv7lxz',
-      yearLearned: 2020,
-      years: 3,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Database',
-      stackArea: 'Data Storage & Databases',
-      stackCategory: 'SQL',
-      name: 'PostgreSQL',
-      icon: '/postgresql.svg',
-      cloudinary: 'postgresql_dx4o4h',
-      yearLearned: 2019,
-      years: 5,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Database',
-      stackArea: 'Data Storage & Databases',
-      stackCategory: 'SQL',
-      name: 'SQL',
-      icon: '/sql.svg',
-      cloudinary: 'sql_dxvo6o',
-      yearLearned: 2007,
-      years: 10,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'APIs',
-      stackArea: 'Backend',
-      stackCategory: 'API Design & Development',
-      name: 'RESTful APIs',
-      icon: '/rest.svg',
-      cloudinary: 'rest_vt3dwy',
-      yearLearned: 2007,
-      years: 12,
-      rating: 9,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'APIs',
-      stackArea: 'Backend',
-      stackCategory: 'API Design & Development',
-      name: 'GraphQL',
-      icon: '/graphql.svg',
-      cloudinary: 'graphql_ztconi',
-      yearLearned: 2019,
-      years: 5,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: true,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Architecture',
-      stackArea: 'Architecture & Patterns',
-      stackCategory: '',
-      name: 'Microservices',
-      icon: '/microservice.svg',
-      cloudinary: 'microservice_xcczrj',
-      yearLearned: 2022,
-      years: 2,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: true,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Architecture',
-      stackArea: 'Architecture & Patterns',
-      stackCategory: '',
-      name: 'Micro-frontends',
-      icon: '/micro-frontend.svg',
-      cloudinary: 'micro-frontend_sx9fbr',
-      yearLearned: 2022,
-      years: 2,
-      rating: 8,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Cloud Platform',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Cloud Platforms',
-      name: 'GCP',
-      icon: '/gcp.svg',
-      cloudinary: 'gcp_h0xhhn',
-      yearLearned: 2021,
-      years: 3,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Cloud Platform',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Cloud Platforms',
-      name: 'GKE',
-      icon: '/gke.svg',
-      cloudinary: 'gke_dwffq2',
-      yearLearned: 2021,
-      years: 3,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Cloud Platform',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Cloud Platforms',
-      name: 'AWS',
-      icon: '/aws.svg',
-      cloudinary: 'aws_ntzjf4',
-      yearLearned: 2018,
-      years: 2,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Containerization',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Containerization & Orchestration',
-      name: 'Docker',
-      icon: '/docker.svg',
-      cloudinary: 'docker_fx1vot',
-      yearLearned: 2021,
-      years: 4,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Containerization',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Containerization & Orchestration',
-      name: 'Colima',
-      icon: '/colima-no-bg.png',
-      cloudinary: 'colima-no-bg_t4awkl',
-      yearLearned: 2023,
-      years: 2,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: false
-   },
-   {
-      type: 'Infrastructure Orchestration',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Containerization & Orchestration',
-      name: 'Kubernetes',
-      icon: '/kubernetes.svg',
-      cloudinary: 'kubernetes_z2x5da',
-      yearLearned: 2021,
-      years: 3,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Infrastructure Orchestration',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Containerization & Orchestration',
-      name: 'Helm',
-      icon: '/helm.svg',
-      cloudinary: 'helm_nqdtj8',
-      yearLearned: 2022,
-      years: 3,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   },
-   {
-      type: 'Infrastructure Orchestration',
-      stackArea: 'DevOps & Cloud Infrastructure',
-      stackCategory: 'Infrastructure as Code',
-      name: 'Terraform',
-      icon: '/terraform.svg',
-      cloudinary: 'terraform_elhgrr',
-      yearLearned: 2022,
-      years: 3,
-      rating: 6,
-      haveUsedProfessionally: true,
-      haveUsedForFun: false,
-      featureOnHome: false,
-      showInSkillsTable: true
-   }
-];
-export const experience = [
-   {
-      role: 'Contract Web Developer',
-      company: 'Self-Employed',
-      location: 'Portland, OR',
-      start: '2024-12',
-      end: '',
-      duties: [
-         {
-            name: '',
-            description:
-               'Building fast, beautiful websites and apps using JavaScript, React, Next.js, Vercel, MongoDB, Node, CSS Modules, and Tailwind CSS. Focused on clean architecture, maintainable code, performance, usability, and accessibility.',
-            skillNames: [
-               'JavaScript',
-               'TypeScript',
-               'Node.js',
-               'React.js',
-               'Next.js',
-               'Redux',
-               'Vercel',
-               'CSS Modules',
-               'CSS',
-               'Tailwind',
-               'RESTful APIs',
-               'Fastify',
-               'MongoDB',
-               'Zustand'
-            ]
-         }
-      ]
-   },
-   {
-      role: 'Staff Full-Stack Software Engineer',
-      company: 'ZenBusiness',
-      location: 'Austin, TX (Remote)',
-      start: '2021-02',
-      end: '2024-12',
-      duties: [
-         {
-            name: 'UI/UX Development & Integration',
-            description:
-               'Built business-critical apps, components, libraries, and design systems using JavaScript, React, Redux, Vue.js, Vuex, CSS Modules, and Styled Components.',
-            moreInfo: [
-               [
-                  'Built real-time, interactive applications with complex state management.',
-                  'Developed and maintained apps, components, libraries, and design systems, enhancing user experience (UX), developer experience (DX), and performance.',
-                  'Implemented responsive design and accessibility best practices.',
-                  'Contributed to the development of customer-facing applications, including onboarding flows, shopping/cards, dashboards, post-purchase/upsells and more.',
-                  'Contributed to the development of internal tools and applications.',
-                  'Improved and maintained the company’s marketing website using PHP, WordPress, JavaScript, CSS, Fastly, Cypress, and SEO, significantly improving core web vitals and coding standards.',
-                  'Led key improvements to performance, accessibility, and usability.',
-                  'Led efforts to improve performance and coding practices for multiple React-based repos.',
-                  'Developed and maintained a design system for consistent UI across products.',
-                  'Collaborated with design teams to implement engaging, pixel-perfect designs.',
-                  'Collaborated with product engineers, UI/UX designers and other cross-functional stakeholders.',
-                  'Developed atomic component libraries for consistent and efficient development.',
-                  'Integrated UI with APIs and data sources.',
-                  'Mentored junior and senior engineers and provided technical guidance.',
-                  'Defined technical requirements, collaborated with Product and Design teams, and led story mapping sessions.',
-                  'Improved coding and logging standards, ensuring clean, maintainable code across projects.'
-               ]
-            ],
-            skillNames: [
-               'JavaScript',
-               'TypeScript',
-               'Node.js',
-               'React.js',
-               'Vue.js',
-               'Redux',
-               'CSS',
-               'Sass/SCSS',
-               'CSS Modules',
-               'Styled Components',
-               'Jest',
-               'Cypress',
-               'RESTful APIs',
-               'A/B Testing',
-               'Multivariate Experimentation',
-               'LaunchDarkly',
-               'PostgreSQL',
-               'MongoDB',
-               'GraphQL',
-               'Micro-frontends',
-            ]
-         },
-         {
-            name: 'Web Development & Optimization',
-            description:
-               "Enhanced the company's public website using WordPress, PHP, SCSS/SASS, and JavaScript. Improved core web vitals and coding standards. Built a component library to increase efficiency and consistency.",
-            moreInfo: [
-               [
-                  'Developed, implemented and documented a component library for consistent UI across the site.',
-                  'Supported the shift to a page-speed-based SEO strategy, collaborating with the SEO Team to meet performance targets (e.g. TTFB, LCP, CLS).',
-                  'Vastly improved core web vitals and coding standards by making systematic optimization and recording performance metrics.',
-                  'Provided documentation, training and support related to maintaining performance and coding best practices.',
-                  'Planned, implemented, documented and supported an atomic component structure, facilitating quick and consistent development.',
-                  'Mentored junior and senior engineers and provided technical guidance.',
-                  'Built automated tests using Cypress.'
-               ]
-            ],
-            // moreInfoLink: {
-            //    href: '/experience/zb-seo',
-            //    text: 'Read more about optimizing the website...'
-            // },
-            skillNames: [
-               'WordPress',
-               'PHP',
-               'Sass/SCSS',
-               'CSS Modules',
-               'CSS',
-               'JavaScript',
-               'Responsive Design',
-               'SEO',
-               'Performance Optimization',
-               'A/B Testing',
-               'Multivariate Experimentation',
-               'LaunchDarkly'
-            ]
-         },
-         {
-            name: 'Microservices & APIs',
-            description:
-               'Built and maintained scalable microservices to enhance functionality and replace monolithic APIs. Used Node.js, Fastify, JavaScript, GCP, Kubernetes, and Docker to develop microservices and APIs.',
-            moreInfo: [
-               [
-                  'Built and maintained scalable microservices to improve developer experience.',
-                  'Leveraged technologies like Node.js, Fastify, JavaScript, GCP...',
-                  'Supported the decommissioning of a monolithic API.',
-                  'Contributed to the development of Node.js/Fastify microservices.',
-                  'Developed robust solutions for validating JSON and XML data, and automatically converting between data formats.',
-                  'Led the development of a critical microservice to automate the gathering, storage, preparation, validation and submission of data for a federal API, boosting efficiency and revenue.',
-                  'Defined requirements, collaborated with stakeholders, led story mapping and development. ',
-                  'Integrated logging and monitoring in GCP to ensure system reliability.',
-                  'Helped improve team coding and logging standards.',
-                  'Built unit tests using Jest.',
-                  'Provided technical support across codebases.',
-                  'Mentored senior and junior engineers.',
-                  'Implemented secure storage and processing of Personally Identifiable Information (PII) using a proxy to handle sensitive data.',
-                  'Provided technical guidance throughout the development lifecycle and represented the team in cross-functional meetings.',
-                  'Managed projects from technical scoping to deployment and user support, ensuring delivery of high-quality features aligned with business goals.'
-               ]
-            ],
-            skillNames: [
-               'JavaScript',
-               'Node.js',
-               'Fastify',
-               'Python',
-               'Jest',
-               'Microservices',
-               'RESTful APIs',
-               'GCP',
-               'GKE',
-               'AWS',
-               'Docker',
-               'Colima',
-               'Kubernetes',
-               'Terraform',
-               'Django',
-               'Micro-frontends',
-            ]
-         },
-         {
-            name: 'Experimentation',
-            description:
-               'Led a team dedicated to experimentation and optimization, focusing on A/B testing and multivariate experimentation. Helped build a culture of experimentation to speed iterative improvements and increase revenue.',
-            moreInfo: [
-               [
-                  'Led planning, building, tracking, and reporting on A/B and multivariate tests.',
-                  'Collaborated with Design and Product teams to develop hypotheses, define metrics, and design variants.',
-                  'Built variants.',
-                  'Worked with Data teams to analyze and present results.',
-                  'Partnered with stakeholders to confirm winners, implement solutions, and plan follow-up tests.',
-                  'Mentored senior engineers on testing and coding practices.',
-                  'Facilitated the transition to LaunchDarkly as our needs became too complex for Google Optimize.',
-                  'Trained other teams and presented to stakeholders regarding experimentation and LaunchDarkly.',
-                  'Helped build a culture of experimentation to speed iterative improvements and increase revenue.',
-                  'Acquired LaunchDarkly Gold Certification and provided ongoing support.'
-               ]
-            ],
-            skillNames: [
-               'A/B Testing',
-               'Multivariate Experimentation',
-               'LaunchDarkly',
-               'JavaScript',
-               'React.js',
-               'Vue.js',
-               'PHP',
-               'CSS',
-               'Sass/SCSS',
-               'Styled Components',
-               'Responsive Design'
-            ]
-         },
-         {
-            name: 'Developer Support & DevOps',
-            description:
-               "Focused on supporting developers and improving developer experience, gained basic DevOps knowledge and experience. While my passions lie closer to the frontend, it's extremely valuable to understand the backend and DevOps.",
-            moreInfo: [
-               [
-                  'Provided technical support and guidance to developers.',
-                  'Administered and documented developer tools.',
-                  'Contributed to API abstractions.',
-                  'Contributed to release workflows and CI/CD.',
-                  'Contributed to a centralized event distributor using Pub/Sub.'
-               ]
-            ],
-            skillNames: [
-               'ArgoCD',
-               'GCP',
-               'Kubernetes',
-               'Pub/Sub',
-               'Auth0',
-               'Twilio',
-               'Mailchimp',
-               'JavaScript',
-               'Node.js',
-               'RESTful APIs',
-               'Docker',
-               'Colima',
-               'Helm',
-               'Terraform'
-            ]
-         }
-      ],
-      skills: [
-         'JavaScript',
-         'TypeScript',
-         'React.js',
-         'Vue.js',
-         'Next.js',
-         'Redux',
-         'Thunk',
-         'CSS',
-         'SCSS',
-         'Sass',
-         'Styled Components',
-         'Responsive Design',
-         'A/B and multivariate testing',
-         'LaunchDarkly',
-         'Google Optimize',
-         'Webpack',
-         'Turbopack',
-         'Parcel',
-         'Jest',
-         'Cypress',
-         'Node.js',
-         'PHP',
-         'WordPress',
-         'Python',
-         'Express',
-         'Fastify',
-         'Django',
-         'MongoDB',
-         'PostgreSQL',
-         'SQL',
-         'RESTful APIs',
-         'GraphQL',
-         'Microservices',
-         'micro-frontends',
-         'GCP',
-         'GKE',
-         'AWS',
-         'Docker',
-         'Colima',
-         'Kubernetes',
-         'Helm',
-         'Terraform'
-      ]
-   },
-   {
-      role: 'Owner & Lead Engineer',
-      company: 'NuTandem',
-      location: 'Portland, OR',
-      start: '2013-05',
-      end: '2021-12',
-      duties: [
-         {
-            name: 'App Development',
-            description:
-               'Developed hundreds of custom apps and websites using JavaScript, React, Node, d3.js, CSS Modules, Styled Components, and Tailwind CSS',
-            moreInfo: [
-               [
-                  'Built a library of reusable, stylable, composable components to speed delivery and facililtate consistent usability, accessability and performance.',
-                  'Oversaw operations. Owned projects from requirements to delivery.'
-               ]
-            ],
-            skillNames: [
-               'JavaScript',
-               'TypeScript',
-               'Node.js',
-               'React.js',
-               'Redux',
-               'D3.js',
-               'Sass/SCSS',
-               'CSS Modules',
-               'Styled Components',
-               'Tailwind',
-               'CSS',
-               'Responsive Design',
-               'RESTful APIs',
-               'GraphQL',
-               'PostgreSQL',
-               'AWS',
-            ]
-         },
-         {
-            name: 'Web Development, SEO & Optimization',
-            description:
-               'Built our company website with WordPress, PHP, JavaScript, and CSS. Optimized SEO and performance.',
-            moreInfo: [
-               [
-                  'Designed and built our company website with WordPress, PHP, JavaScript, and CSS, including custom templates, menus, and widgets, to facilitate frequent updates.',
-                  'Led SEO, marketing and sales efforts.',
-                  'Managed the company’s online presence and digital marketing strategy.',
-                  'Implemented SEO best practices to improve search engine rankings and drive organic traffic.'
-               ]
-            ],
-            skillNames: [
-               'Next.js',
-               'WordPress',
-               'PHP',
-               'JavaScript',
-               'CSS Modules',
-               'CSS',
-               'SEO',
-               'Performance Optimization',
-               'Responsive Design'
-            ]
-         },
-         {
-            name: 'Everything Else',
-            description:
-               'Established and maintained relationships with over 50 organizations including state and federal agencies, restaurants like Sizzler and Outback, universities, and large corporations like Unilever.',
-            moreInfo: [
-               [
-                  'Managed all aspects of the business, including relationships, marketing, and customer support.',
-                  'Maintained solid relationships with large corporations (like Massage Envy, Sizzler Restaurants, Outback Steakhouse, and Unilever), government agencies (such as the USDA Forest Service and multiple state agencies), universities (including University of Montana, Central Oregon Community College, and Carroll University) and a variety of other customers.',
-                  'Provided training and support to clients.',
-                  'Owned projects from discovery to delivery.',
-                  'Conducted usability studies and worked with clients to understand their unique needs.'
-               ]
-            ]
-         }
-      ]
-   },
-   {
-      role: 'Contract Web Developer',
-      company: 'USDA Forest Service',
-      location: 'Albuquerque, NM (Remote)',
-      start: '2013-06',
-      end: '2020-12',
-      duties: [
-         {
-            name: 'Web Development & Performance Optimization',
-            description:
-               'Managed web projects and improved system performance using JavaScript, CSS, PHP and SQLite.',
-            moreInfo: [
-               [
-                  'Developed and maintained a custom web application for the USDA Forest Service.',
-                  'Juggled a steady stream of web projects for internal websites and the public website at www.fs.usda.gov',
-                  'Built web apps with JavaScript, CSS, PHP, and SQLite.',
-                  'Reduced reliance on jQuery by implementing pure JavaScript components.',
-                  'Built templates to enable self-serve website creation for non-technical users.',
-                  'Used Git and GitHub to share, review, and manage code. Participated in code reviews and incremental refactoring.',
-                  'Helped demonstrate app enhancements and performance improvements to leadership.',
-                  "Lead a project to build a Content Review Tool to support the agency's cumbersome efforts to review and improve hundreds of thousands of pages in their  public website."
-               ]
-            ],
-            skillNames: [
-               'JavaScript',
-               'Node.js',
-               'PHP',
-               'SQLite',
-               'SQL',
-               'CSS',
-               'CSS3',
-               'Responsive Design',
-               'Performance Optimization',
-               'RESTful APIs'
-            ]
-         },
-         {
-            name: 'Web Accessibility & Training Expert',
-            description:
-               'Trained colleagues, conducted code reviews, and managed web projects. Built web apps using JavaScript, CSS, PHP and SQLite.',
-            moreInfo: [
-               [
-                  'Trained with a senior web accessbility expert (WCAG).',
-                  'Improved web accessibility for Forest Service websites and templates.',
-                  'Designed and delivered WCAG training.',
-                  'Provided training related to self-serve website creation.',
-                  'Provided ongoing technical support.'
-               ]
-            ],
-            skillNames: ['Accessibility (WCAG)']
-         }
-      ]
-   },
-   {
-      role: 'Systems Manager & Process Engineer',
-      company: 'GL Solutions',
-      location: 'Bend, OR',
-      start: '2006-07',
-      end: '2013-05',
-      duties: [
-         {
-            name: 'Systems Manager & Process Engineer',
-            description:
-               'Re-engineered processes and tools across the organization. Improved efficiency, accuracy, customer satisfaction and employee engagement, and usability and adoption for internal software tools. Defined SCAS, KPIs, and team metrics. Implemented a continuous improvement initiative, empowering employees to actively participate in evolving the organization. Provided custom software, documentation, training, and support.',
-            moreInfo: [
-               [
-                  'Re-engineered processes and tools across the organization.',
-                  'Improved efficiency, accuracy, customer satisfaction and employee engagement.',
-                  'Defined SCAS, KPIs, and team metrics.',
-                  'Implemented a continuous improvement initiative, empowering employees to actively participate in evolving the organization.',
-                  'Provided custom software, documentation, training, and support.',
-                  'Integrated customer service, sales, and marketing teams into a cohesive Business Solutions Department.',
-                  'Re-engineered processes and tools to better utilize talent and experience.',
-                  'Worked directly with customers to increase engagement.',
-                  'Facilitated idea sharing and built a customer community.',
-                  'Conducted stakeholder and SME interviews, documented requirements, and produced software specifications.',
-                  'Authored and maintained technical content, and improved documentation standards.',
-                  'Wrote detailed responses to government RFPs, RFQs and RFIs. Maintained an in-depth library of content for use in formal responses. Responded to questions and provided support to sales to close deals and transition new customers.',
-                  'Conducted stakeholder and SME interviews, documented requirements, and produced software specifications. Authored and maintained technical content, and improved documentation standards.'
-               ]
-            ],
-            skillNames: [
-               'SQL',
-               'CSS',
-               'HTML5',
-               'Responsive Design',
-               'Usability Testing',
-               'User Experience (UX)',
-               'User Interface (UI)',
-               'Agile Development',
-               'Scrum'
-            ]
-         }
-      ]
-   },
-   {
-      role: 'Technical Documentation Specialist',
-      company: 'I/O Software',
-      location: 'Riverside, CA',
-      start: '2002-03',
-      end: '2006-07',
-      duties: [
-         {
-            name: '',
-            description:
-               'Produced technical documentation for end users and internal users. Contributed to patent applications related to cryptographic security.',
-            moreInfo: [
-               [
-                  'Produced user guides, in-app help systems, and technical articles related to software integrations with biometric security devices (e.g. fingerprint, retina, iris, voice, face, and keystroke).',
-                  'Interviewed developers and other stakeholders.',
-                  'Contributed to cryptographic patent applications related to security algorithms and biometric data mappings, including research, mathematical consultation and diagrams.'
-               ]
-            ],
-            skillNames: ['CSS', 'HTML5']
-         }
-      ]
-   }
-];
 export const projects = [
    {
       name: 'Things App v2.0',
@@ -2485,75 +1268,6 @@ export const projects = [
       featureOnHome: false
    }
 ];
-export const education = [
-   {
-      degree: 'Bachelor of Science',
-      major: 'Mathematics',
-      institution: 'University of Redlands',
-      location: 'Redlands, CA',
-      startYear: 1996,
-      endYear: 2001,
-      awards: [
-         'Endowed Alumni Scholarship',
-         'Presidential Scholarship',
-         'Scholar-Athlete'
-      ]
-   },
-   {
-      degree: 'Master-Level Coursework',
-      major: 'Applied Mathematics',
-      institution: 'Claremont Graduate University',
-      location: 'Claremont, CA',
-      startYear: 2001,
-      endYear: 2001,
-      awards: [
-         'Fully Paid PhD Fellowship',
-         'Excelled, tutored peers, led study groups',
-         'Focused on fluid dynamics and differential equations',
-         'Participated in sponsored workshops to optimize semiconductor performance',
-         'Helped distinguished professors emeritus produce the second edition of their popular textbook on differential equations. Devised complex mathematical problems and authored solutions. Edited and proofread the text. Used LaTeX to typeset the book. Trained other students in LaTeX. Developed a love of tech. Also enhanced my life-long love of mentoring and teaching.',
-         "Wasn't planning on grad school but couldn't pass up the fully-paid PhD fellowship. Although I enjoyed the work, I realized that I didn't want to be a mathematician. I followed my new-found love of software to a career in tech."
-      ]
-   }
-];
-export const certifications = [
-   {
-      name: 'LaunchDarkly Gold Certification',
-      date: '2024'
-   },
-   {
-      name: 'Software Security Training',
-      date: '2021-2024'
-   },
-   {
-      name: 'Information Security Awareness (AgLearn)',
-      date: '2020'
-   },
-   {
-      name: 'Nielsen Norman Group User Experience training',
-      date: '2012'
-   }
-];
-
-// todo: attach as methods/properties to each data object
-
-export const homeSkills = skills
-   .filter(skill => skill.featureOnHome === true)
-   .map(skill => skill.name);
-
-export const ctaSkills = skills
-   .filter(skill => skill.featureInCTAs === true)
-   .map(skill => skill.name);
-// .sort((a, b) => b.rating - a.rating);
-
-export const skillsForTable = skills
-   .filter(skill => skill.showInSkillsTable === true)
-   .sort((a, b) => {
-      if (b.rating !== a.rating) {
-         return b.rating - a.rating; // Sort by rating first
-      }
-      return b.years - a.years; // If ratings are equal, sort by years
-   });
 
 export const homeProjects = projects.filter(
    project => project.featureOnHome === true
@@ -2563,59 +1277,36 @@ export const projectsForGallery = projects.filter(
    project => project.showInProjectsGallery === true
 );
 
-const filterAndInverse = (array, callback) => {
+type FilterCallback = (item: DefinitionProject) => boolean;
+
+const filterAndInverse = (
+   array: DefinitionProject[],
+   callback: FilterCallback
+): [DefinitionProject[], DefinitionProject[]] => {
    const filtered = array.filter(callback);
    const inverseFiltered = array.filter(item => !callback(item));
    return [filtered, inverseFiltered];
 };
+
 export const [recentProjects, olderProjects] = filterAndInverse(
    projectsForGallery,
    project => getYearsSince(project.year) < 2
 );
 
-const getFeaturedImage = function () {
-   // con't be arrow function because it needs to be able to access 'this'
-   return this.images.find(image => image.type !== 'video');
+const getFeaturedImage = function (
+   project: DefinitionProject
+): ProjectImage | undefined {
+   if (!project.images) return undefined;
+   return project.images.find((image: ProjectImage) => image.type !== 'video');
 };
 
-const getModalImages = function () {
-   return this.images.filter(image => !image.hideFromModal);
+const getModalImages = function (project: DefinitionProject) {
+   if (!project.images) return undefined;
+   return project.images.filter((image: ProjectImage) => !image.hideFromModal);
 };
 
-projects.forEach(project => {
-   project.getFeaturedImage = getFeaturedImage;
-   project.getModalImages = getModalImages;
+projects.forEach((project: DefinitionProject) => {
+   project.getFeaturedImage = (): ProjectImage | undefined =>
+      getFeaturedImage(project);
+   project.getModalImages = () => getModalImages(project);
 });
-
-const yearsSinceLearned = function () {
-   const yearWithMonth = `${this.yearLearned}-01`;
-   return getYearsSince(yearWithMonth);
-};
-
-function getJobsWithSkill(skillName) {
-   return experience.filter(job =>
-      job.duties.some(duty => duty.skillNames?.includes(skillName))
-   );
-}
-
-function getProjectsWithSkill(skillName) {
-   return projects.filter(project => project.tech?.includes(skillName));
-}
-
-skills.forEach(skill => {
-   skill.yearsSinceLearned = yearsSinceLearned;
-   skill.getJobsWithSkill = () => getJobsWithSkill(skill.name); // Use as a shared function
-   skill.getProjectsWithSkill = () => getProjectsWithSkill(skill.name);
-});
-
-export const skillIcon = skills.map(skill => skill.cloudinary);
-export const skillIcons = (skillNames = [], dark = false) => {
-   return skills
-      .filter(skill => skillNames.includes(skill.name))
-      .map(skill => ({
-         src: dark
-            ? skill.cloudinaryDark || skill.cloudinary
-            : skill.cloudinary,
-         altText: skill.name
-      }));
-};
