+++ data/skills.ts
@@ -1,0 +1,791 @@
+import { experience } from '@/data/experience';
+import { projects } from '@/data/projects';
+import { getYearsSince } from '@/utils/date';
+import { DefinitionSkill } from '@/types/Skill';
+
+export const skills: DefinitionSkill[] = [
+   {
+      type: 'Language',
+      stackArea: 'General',
+      stackCategory: 'Language',
+      name: 'JavaScript',
+      icon: '/js.svg',
+      cloudinary: 'js_au9kmj',
+      cloudinaryDark: 'javascript-dark-icon_diwklq',
+      yearLearned: 2013,
+      years: 12,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Frontend Library',
+      stackArea: 'Frontend',
+      stackCategory: 'Library',
+      name: 'React.js',
+      icon: '/react.svg',
+      cloudinary: 'react_g4rwdx',
+      cloudinaryDark: 'react-dark-icon_dibccl',
+      yearLearned: 2015,
+      years: 9,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Frontend Framework',
+      stackArea: 'Frontend',
+      stackCategory: 'Framework',
+      name: 'Next.js',
+      icon: '/next.svg',
+      cloudinary: 'nextjs_jxdbqq',
+      cloudinaryDark: 'next-dark-icon_qfvtkk',
+      yearLearned: 2019,
+      years: 4,
+      rating: 9,
+      haveUsedProfessionally: false,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Frontend Framework',
+      stackArea: 'Frontend',
+      stackCategory: 'Framework',
+      name: 'Vue.js',
+      icon: '/vue.svg',
+      cloudinary: 'vue_thuhnk',
+      cloudinaryDark: 'vue-dark-icon_xghpvp',
+      yearLearned: 2021,
+      years: 3,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Backend Development',
+      stackArea: 'Backend',
+      stackCategory: 'Language',
+      name: 'Node.js',
+      icon: '/node.svg',
+      cloudinary: 'node_hl58om',
+      cloudinaryDark: 'node-dark-icon_kl7oy6',
+      yearLearned: 2015,
+      years: 9,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Language',
+      stackArea: 'General',
+      stackCategory: 'Language',
+      name: 'TypeScript',
+      icon: '/ts1.svg',
+      cloudinary: 'ts1_nbdvxm',
+      cloudinaryDark: 'typescript-dark-icon_mj51g8',
+      yearLearned: 2015,
+      years: 9,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'State Management',
+      stackArea: 'Frontend',
+      stackCategory: 'Library',
+      name: 'Redux',
+      icon: '/redux.svg',
+      cloudinary: 'redux_vd28jd',
+      yearLearned: 2019,
+      years: 8,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'State Management',
+      stackArea: 'Frontend',
+      stackCategory: 'Library',
+      name: 'Zustand',
+      icon: '/react.svg',
+      cloudinary: 'zustand_wrx1q3',
+      yearLearned: 2024,
+      years: 1,
+      rating: 8,
+      haveUsedProfessionally: false,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Visualization',
+      stackArea: 'Frontend',
+      stackCategory: 'Library',
+      name: 'D3.js',
+      cloudinary: 'd3js_eakt6f',
+      yearLearned: 2013,
+      years: 8,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Visualization',
+      stackArea: 'Frontend',
+      stackCategory: 'Library',
+      name: 'React Simple Maps',
+      cloudinary: 'react-simple-maps_xvvujp',
+      yearLearned: 2013,
+      years: 8,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Styling',
+      stackArea: 'Frontend',
+      stackCategory: 'Styling',
+      name: 'CSS',
+      icon: '/css3.svg',
+      cloudinary: 'css3_mvgbbf',
+      yearLearned: 2004,
+      years: 12,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Styling',
+      stackArea: 'Frontend',
+      stackCategory: 'Styling',
+      name: 'CSS Modules',
+      icon: '/sass.svg',
+      cloudinary: 'css3_mvgbbf',
+      cloudinaryDark: 'css3-dark-icon_xtdb3t',
+      yearLearned: 2015,
+      years: 8,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Styling',
+      stackArea: 'Frontend',
+      stackCategory: 'Styling',
+      name: 'Sass/SCSS',
+      icon: '/sass.svg',
+      cloudinary: 'sass_s3kz3x',
+      cloudinaryDark: 'sass-dark-icon_h3esvt',
+      yearLearned: 2015,
+      years: 8,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Styling',
+      stackArea: 'Frontend',
+      stackCategory: 'Styling',
+      name: 'Tailwind',
+      icon: '/tailwind.svg',
+      cloudinary: 'tailwind_k9r34s',
+      cloudinaryDark: 'tailwind-dark-icon_mbvvtr',
+      yearLearned: 2019,
+      years: 4,
+      rating: 9,
+      haveUsedProfessionally: false,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Styling',
+      stackArea: 'Frontend',
+      stackCategory: 'Styling',
+      name: 'JSS',
+      icon: '/jss.svg',
+      cloudinary: 'jss_zewece',
+      yearLearned: 2016,
+      years: 8,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: false
+   },
+   {
+      type: 'Styling',
+      stackArea: 'Frontend',
+      stackCategory: 'Styling',
+      name: 'Styled Components',
+      icon: '/styled-components.svg',
+      cloudinary: 'styled-components_fd8fvo',
+      cloudinaryDark: 'styled-components-dark-icon_fs0kmr',
+      yearLearned: 2016,
+      years: 8,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Styling',
+      stackArea: 'Frontend',
+      stackCategory: 'Styling',
+      name: 'Panda CSS',
+      icon: '/react.svg',
+      cloudinary: 'panda-css_av1hkt',
+      yearLearned: 2025,
+      years: 1,
+      rating: 9,
+      haveUsedProfessionally: false,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: false
+   },
+   {
+      type: 'UI/UX',
+      stackArea: 'Frontend',
+      stackCategory: 'UI/UX',
+      name: 'Responsive Design',
+      icon: '/rwd.svg',
+      cloudinary: 'rwd_o9z8s0',
+      yearLearned: 2007,
+      years: 11,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'UI/UX',
+      stackArea: 'Experimentation & Optiminzation',
+      stackCategory: '',
+      name: 'A/B Testing',
+      icon: '/beaker.svg',
+      cloudinary: 'beaker_vqshpg',
+      yearLearned: 2010,
+      years: 3,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'UI/UX',
+      stackArea: 'Experimentation & Optiminzation',
+      stackCategory: '',
+      name: 'Multivariate Experimentation',
+      icon: '/beaker1.svg',
+      cloudinary: 'beakers_yellow_eehkpx',
+      yearLearned: 2022,
+      years: 3,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'UI/UX',
+      stackArea: 'Experimentation & Optiminzation',
+      stackCategory: '',
+      name: 'LaunchDarkly',
+      icon: '/launchdarkly.svg',
+      cloudinary: 'launchdarkly_kdznir',
+      yearLearned: 2022,
+      years: 3,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Build Tools',
+      stackArea: 'Build & Testing Tools',
+      stackCategory: 'Build Tools',
+      name: 'Webpack',
+      icon: '/webpack.svg',
+      cloudinary: 'webpack_ycdnkv',
+      yearLearned: 2015,
+      years: 7,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Build Tools',
+      stackArea: 'Build & Testing Tools',
+      stackCategory: 'Build Tools',
+      name: 'Turbopack',
+      icon: '/turbopack.svg',
+      cloudinary: 'turbopack_gvkvbu',
+      yearLearned: 2024,
+      years: 1,
+      rating: 7,
+      haveUsedProfessionally: false,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Build Tools',
+      stackArea: 'Build & Testing Tools',
+      stackCategory: 'Build Tools',
+      name: 'Parcel',
+      icon: '/parcel.svg',
+      cloudinary: 'parcel_ik2cum',
+      yearLearned: 2024,
+      years: 1,
+      rating: 7,
+      haveUsedProfessionally: false,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: false
+   },
+   {
+      type: 'Testing',
+      stackArea: 'Build & Testing Tools',
+      stackCategory: 'Testing',
+      name: 'Jest',
+      icon: '/jest.svg',
+      cloudinary: 'jest_djvbcr',
+      yearLearned: 2022,
+      years: 3,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Testing',
+      stackArea: 'Build & Testing Tools',
+      stackCategory: 'Testing',
+      name: 'Cypress',
+      icon: '/cypress.svg',
+      cloudinary: 'cypress_ihahr3',
+      yearLearned: 2021,
+      years: 1,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Backend Development',
+      stackArea: 'Backend',
+      stackCategory: 'Language',
+      name: 'PHP',
+      icon: '/php.svg',
+      cloudinary: 'php-icon_apfbnf',
+      cloudinaryDark: 'php01-dark-icon_fzefry',
+      yearLearned: 2013,
+      years: 9,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Backend Development',
+      stackArea: 'General',
+      stackCategory: 'Content Management',
+      name: 'WordPress',
+      icon: '/wordpress.svg',
+      cloudinary: 'wordpress_xkgfcd',
+      cloudinaryDark: 'wordpress-dark-icon_volzbh',
+      yearLearned: 2013,
+      years: 9,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Backend Development',
+      stackArea: 'Backend',
+      stackCategory: 'Language',
+      name: 'Python',
+      icon: '/python.svg',
+      cloudinary: 'python_g5utim',
+      cloudinaryDark: 'python-dark-icon_ogismc',
+      yearLearned: 2021,
+      years: 3,
+      rating: 7,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: true,
+      showInSkillsTable: true,
+      featureInCTAs: true
+   },
+   {
+      type: 'Backend Framework',
+      stackArea: 'Backend',
+      stackCategory: 'Framework',
+      name: 'Express',
+      icon: '/express.svg',
+      cloudinary: 'express_bqzgin',
+      yearLearned: 2015,
+      years: 9,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Backend Framework',
+      stackArea: 'Backend',
+      stackCategory: 'Framework',
+      name: 'Fastify',
+      icon: '/fastify.svg',
+      cloudinary: 'fastify_zy8jul',
+      yearLearned: 2023,
+      years: 2,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Backend Framework',
+      stackArea: 'Backend',
+      stackCategory: 'Framework',
+      name: 'Django',
+      icon: '/django.svg',
+      cloudinary: 'django_yilgyq',
+      yearLearned: 2020,
+      years: 2,
+      rating: 4,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Database',
+      stackArea: 'Data Storage & Databases',
+      stackCategory: 'NoSQL',
+      name: 'MongoDB',
+      icon: '/mongodb.svg',
+      cloudinary: 'mongodb_kv7lxz',
+      yearLearned: 2020,
+      years: 3,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Database',
+      stackArea: 'Data Storage & Databases',
+      stackCategory: 'SQL',
+      name: 'PostgreSQL',
+      icon: '/postgresql.svg',
+      cloudinary: 'postgresql_dx4o4h',
+      yearLearned: 2019,
+      years: 5,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Database',
+      stackArea: 'Data Storage & Databases',
+      stackCategory: 'SQL',
+      name: 'SQL',
+      icon: '/sql.svg',
+      cloudinary: 'sql_dxvo6o',
+      yearLearned: 2007,
+      years: 10,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'APIs',
+      stackArea: 'Backend',
+      stackCategory: 'API Design & Development',
+      name: 'RESTful APIs',
+      icon: '/rest.svg',
+      cloudinary: 'rest_vt3dwy',
+      yearLearned: 2007,
+      years: 12,
+      rating: 9,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'APIs',
+      stackArea: 'Backend',
+      stackCategory: 'API Design & Development',
+      name: 'GraphQL',
+      icon: '/graphql.svg',
+      cloudinary: 'graphql_ztconi',
+      yearLearned: 2019,
+      years: 5,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: true,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Architecture',
+      stackArea: 'Architecture & Patterns',
+      stackCategory: '',
+      name: 'Microservices',
+      icon: '/microservice.svg',
+      cloudinary: 'microservice_xcczrj',
+      yearLearned: 2022,
+      years: 2,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: true,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Architecture',
+      stackArea: 'Architecture & Patterns',
+      stackCategory: '',
+      name: 'Micro-frontends',
+      icon: '/micro-frontend.svg',
+      cloudinary: 'micro-frontend_sx9fbr',
+      yearLearned: 2022,
+      years: 2,
+      rating: 8,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Cloud Platform',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Cloud Platforms',
+      name: 'GCP',
+      icon: '/gcp.svg',
+      cloudinary: 'gcp_h0xhhn',
+      yearLearned: 2021,
+      years: 3,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Cloud Platform',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Cloud Platforms',
+      name: 'GKE',
+      icon: '/gke.svg',
+      cloudinary: 'gke_dwffq2',
+      yearLearned: 2021,
+      years: 3,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Cloud Platform',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Cloud Platforms',
+      name: 'AWS',
+      icon: '/aws.svg',
+      cloudinary: 'aws_ntzjf4',
+      yearLearned: 2018,
+      years: 2,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Containerization',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Containerization & Orchestration',
+      name: 'Docker',
+      icon: '/docker.svg',
+      cloudinary: 'docker_fx1vot',
+      yearLearned: 2021,
+      years: 4,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Containerization',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Containerization & Orchestration',
+      name: 'Colima',
+      icon: '/colima-no-bg.png',
+      cloudinary: 'colima-no-bg_t4awkl',
+      yearLearned: 2023,
+      years: 2,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: false
+   },
+   {
+      type: 'Infrastructure Orchestration',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Containerization & Orchestration',
+      name: 'Kubernetes',
+      icon: '/kubernetes.svg',
+      cloudinary: 'kubernetes_z2x5da',
+      yearLearned: 2021,
+      years: 3,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Infrastructure Orchestration',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Containerization & Orchestration',
+      name: 'Helm',
+      icon: '/helm.svg',
+      cloudinary: 'helm_nqdtj8',
+      yearLearned: 2022,
+      years: 3,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   },
+   {
+      type: 'Infrastructure Orchestration',
+      stackArea: 'DevOps & Cloud Infrastructure',
+      stackCategory: 'Infrastructure as Code',
+      name: 'Terraform',
+      icon: '/terraform.svg',
+      cloudinary: 'terraform_elhgrr',
+      yearLearned: 2022,
+      years: 3,
+      rating: 6,
+      haveUsedProfessionally: true,
+      haveUsedForFun: false,
+      featureOnHome: false,
+      showInSkillsTable: true
+   }
+];
+
+export const homeSkills = skills
+   .filter(skill => skill.featureOnHome === true)
+   .map(skill => skill.name);
+
+export const ctaSkills = skills
+   .filter(skill => skill.featureInCTAs === true)
+   .map(skill => skill.name);
+
+export const skillsForTable = skills
+   .filter(skill => skill.showInSkillsTable === true)
+   .sort((a, b) => {
+      if (a.rating && b.rating && b.rating !== a.rating) {
+         return b.rating - a.rating; // Sort by rating first
+      }
+      if (a.years && b.years && b.years !== a.years) {
+         return b.years - a.years; // If ratings are equal, sort by years
+      }
+      return 1;
+   });
+
+const yearsSinceLearned = function (yearLearned: number) {
+   const yearWithMonth = `${yearLearned}-01`;
+   return getYearsSince(yearWithMonth);
+};
+
+function getJobsWithSkill(skillName: string) {
+   return experience.filter(job =>
+      job.duties.some(duty => duty.skillNames?.includes(skillName))
+   );
+}
+
+function getProjectsWithSkill(skillName: string) {
+   return projects.filter(project => project.tech?.includes(skillName));
+}
+
+skills.forEach(skill => {
+   skill.yearsSinceLearned = skill.yearLearned
+      ? yearsSinceLearned(skill.yearLearned)
+      : undefined;
+   skill.getJobsWithSkill = () => getJobsWithSkill(skill.name); // Use as a shared function
+   skill.getProjectsWithSkill = () => getProjectsWithSkill(skill.name);
+});
+
+export const skillIcon = skills.map(skill => skill.cloudinary);
+export const skillIcons = (skillNames: string[] = [], dark = false) => {
+   return skills
+      .filter(skill => skillNames.includes(skill.name))
+      .map(skill => ({
+         src: dark
+            ? skill.cloudinaryDark || skill.cloudinary
+            : skill.cloudinary,
+         altText: skill.name
+      }));
+};
