+++ types/Project.ts
@@ -4,17 +4,28 @@ export interface BaseProject {
    type?: string;
 }
 
+export interface ProjectImage {
+   name: string;
+   caption?: string;
+   type?: string;
+   hideFromModal?: boolean;
+   frame?: number;
+}
+
 export interface ModalProject extends BaseProject {
-    description?: string;
-    tech?: string[];
-    liveLink?: string;
-    codeLink?: string;
-    moreInfoLink?: string;
-    motivation?: string;
-    outcome?: string;
-    getModalImages?: () => Array<{
-       name: string;
-       caption: string;
-       type: 'image' | 'video';
-    }>;
- }
+   description?: string;
+   tech?: string[];
+   liveLink?: string;
+   codeLink?: string;
+   moreInfoLink?: string;
+   motivation?: string;
+   outcome?: string;
+   getModalImages?: () => ProjectImage[] | undefined;
+}
+
+export interface DefinitionProject extends ModalProject {
+   images?: ProjectImage[];
+   showInProjectsGallery?: boolean;
+   featureOnHome?: boolean;
+   getFeaturedImage?: () => ProjectImage | undefined;
+}
