+++ types/Skill.ts
@@ -13,3 +13,19 @@ export interface ModalSkill extends BaseSkill {
    getProjectsWithSkill?: () => Project[];
    getJobsWithSkill?: () => Job[];
 }
+
+export interface DefinitionSkill extends BaseSkill {
+   stackArea?: string;
+   stackCategory?: string;
+   icon?: string;
+   cloudinaryDark?: string;
+   yearLearned?: number;
+   haveUsedProfessionally: boolean;
+   haveUsedForFun: boolean;
+   featureOnHome: boolean;
+   showInSkillsTable: boolean;
+   featureInCTAs?: boolean;
+   yearsSinceLearned?: number;
+   getProjectsWithSkill?: () => Project[];
+   getJobsWithSkill?: () => Job[];
+}
